Enhancement: Present user origin on the list of users (WA-359)
Present user **origin** on the list of users:
![image](/uploads/fbc7b9eb8e38b7b1134cde3f0234a164/image.png)

I removed an image (SENS logo) that was presented next to the origin value as we do not have a relevant image for a *GNS* value.
In MVP (v3.4) it's not required.

Closes WA-359

Reviewed-by: Marek Mastyło <mmastylo@silenteight.com>
Reviewed-by: Kamil Dzięcioł <kdzieciol@silenteight.com>
Merged from: https://gitlab.silenteight.com/sens/sens-webapp/merge_requests/102