Bugfix: Consider system property when setting home directory (DO-117)
Previously only environmental variable was considered when
discovering home directory. It is more standard to set the home
directory via system property when building container image
(to not pollute the environment).

On the other hand, when the system property was set, the discovered
home directory was not taken into account and silently
replaced with the system property. This inverted standard
Spring hierarchy of env vars -> system properties -> arguments,
therefore is considered a bug.

Closes DO-117

Reviewed-by: Szymon Pacanowski <spacanowski@silenteight.com>
Reviewed-by: Tadeusz Kleszcz <tkleszcz@silenteight.com>
Merged from: https://gitlab.silenteight.com/sep/sep-base/-/merge_requests/9