Bugfix: Reasoning Branch report displays updatedAt: 1970-01-01 if branch is unmodified (SENS-85-1)
Dates generated in CSV are not aligned with what we display in UI.

The ResoningBranchReporter was always putting 'updated_at' timestamp in CSV/updated_at column. This means that if the reasoning branch has not been persisted (because it corresponds to a brand new feature vector) or has been created but it has not been modified, the updated_at was null which was serialized as 1970-01-01 in the CSV report.

This fix aligns the behaviour with the UI: in the report we put 'update_at'; if it's not present we put 'created at'; if none of them are present we put empty value (corresponding to N/A value displayed in UI).

Reported by the customer under SD-330.

Closes SENS-85

Reviewed-by: Marek Mastyło <mmastylo@silenteight.com>
Reviewed-by: Jacek Obarymski <jobarymski@silenteight.com>
Merged from: https://gitlab.silenteight.com/sens/sens-webapp/-/merge_requests/553

(cherry picked from commit a7c7682e5b62759233d72592c58827317176b851)

Reviewed-by: Kamil Dzięcioł <kdzieciol@silenteight.com>
Reviewed-by: Jacek Obarymski <jobarymski@silenteight.com>
Merged from: https://gitlab.silenteight.com/sens/sens-webapp/-/merge_requests/555