Implement creating matches use case (CS-329)
This patch implements AlertService methods to create matches for alerts.

The sort index in the incoming match is treated literally if set,
or is set to the next one after the max sort index of other matches in
the alert.

Validation of duplicated sort index of incoming matches is done manually,
because it would be hard to do otherwise.

![image](/uploads/10cda475178aacb616a523bd75379890/image.png)

Reviewed-by: Tomasz Bożek <tbozek@silenteight.com>
Reviewed-by: Tadeusz Kleszcz <tkleszcz@silenteight.com>
Merged from: https://gitlab.silenteight.com/sens/adjudication-engine/-/merge_requests/24