Improve comment context
Renaming AlertTemplateModel to AlertContext corresponds with the naming convention of Pebble, which evaluates the template with a context.

After requirements gathering it became apparent that Match context needs to provide a reason for the solution, that comes from a call to Governance, for the Data Science team to use it when preparing templates.

This patch also fixes and refactors a little the infrastructure used to generate comments with Pebble.

Merged from: https://gitlab.silenteight.com/sens/adjudication-engine/-/merge_requests/72