fix(analysis): Serialize executing SelectMissingMatchFeatureQuery (AEP-209)
Due to the message flow, this query might be executed for the same analysis
multiple times in parallel, with different results, duplicating requests
to agents.

The duplication causes that multiple duplicated values come back from agents,
which result in conflicts on INSERTs, potentially missing the results for
differences.

This is just first try to hide the problem, by using advisory lock before
executing the query.