Bugfix: Add gRPC timeout (WEB-1614)
At HSBC Warehouse was not properly registered as gRPC server. As a result Gov's attempt to obtain gRPC client failed, but because there was no timeout set, Gov was waiting indefinitely to acquire the client. This resulted in blocking the scheduler's thread and, as a result, blocked an entire QA.

Closes WEB-1614

Reviewed-by: Przemysław Lewandowski <plewandowski@silenteight.com>
Reviewed-by: Rafał Pajdak <rpajdak@silenteight.com>
Merged from: https://gitlab.silenteight.com/sens/serp-governance/-/merge_requests/210