BUGFIX: Fix error with updating batch status when alert json is invalid (ALL-323)
Closes ALL-323

When some alert in JSON is invalid (but it is not the first alert in bulk) there is an error when updating bulk in the database. Because of that bulk doesn't change status to ERROR. This issue is related to clearing cache by `EntityManager`. I manage to fix that just by simply getting bulk from the database just before changing status and updating.

Memory usage on 14k alerts where some of the last alerts are invalid

![image](/uploads/d77809bafaed11f7c662bb8444c6996e/image.png)

Reviewed-by: Adrian Pozorski <apozorski@silenteight.com>
Reviewed-by: Adam Wójcicki <awojcicki@silenteight.com>
Merged from: https://gitlab.silenteight.com/hsbc-bank/hsbc-bridge/-/merge_requests/351