fix(analysis): Stop recording excessive match features in exchange (AEP-244)
Failure to keep `features` and `agentConfigFeatureIds` in sync results in
excessive records being added to `ae_agent_exchange_match_feature`.

This might not look like too much of a deal, except that we're potentially
recording exchanges being sent for `agent_config_feature_id` which is actually
never sent! This forgoes all attempts of sending the request for feature values
for some matches.